# Conflicts:
#	Sources/Internal/Engine/Private/Qt/PlatformCoreQt.cpp
#	Sources/Internal/Engine/Private/Qt/WindowBackendQt.cpp


Former-commit-id: 55714e1c6989704644eb0ddc2b7031ad653e40a1

+++ Sources/Internal/Network/Services/MMNet/MMBigDataTransferService.cpp
@@ -201,7 +201,7 @@ bool MMBigDataTransferService::BeginNextSnapshot(SnapshotInfo* snapshot)
     fileHandle = File::Create(snapshot->filename, File::OPEN | File::READ);
     if (fileHandle != nullptr)
     {
-        snapshot->fileSize = fileHandle->GetSize();
+        snapshot->fileSize = static_cast<uint32>(fileHandle->GetSize());
         if (snapshot->fileSize > 0)
         {
             SendNextChunk(snapshot);
